+++ lib/statistrano/config.rb
@@ -5,24 +5,41 @@ module Statistrano
   class Config
     include RakeTaskWithContextCreation
 
-    attr_reader :options
-    attr_reader :tasks
+    attr_reader :options,
+                :tasks,
+                :validators
 
     # initalize with the potential for seed options
     # this is required so that when config'd classes
     # are extended we can pass that configuration along
-    def initialize options: nil, tasks: nil
-      @options = options.nil? ? {} : options.clone
-      @tasks   = tasks.nil?   ? {} : tasks.clone
+    def initialize options: nil, tasks: nil, validators: nil
+      @options    = options.nil?    ? {} : options.clone
+      @tasks      = tasks.nil?      ? {} : tasks.clone
+      @validators = validators.nil? ? {} : validators.clone
 
       @options.each do |key,val|
         define_option_accessor key.to_sym
+      end
+
+      @validators.each do |key,val|
+        define_validator key.to_sym
       end
     end
 
     private
 
+      def define_validator name
+        define_singleton_method(:"validator_for_#{name}") do |key, proc, message|
+          @validators[name] = { validator: proc,
+                                message: message }
+        end
 
+        define_singleton_method(:"validate_#{name}") do |key, arg|
+          if @validators.has_key?(name) && !@validators[name][:validator].call(arg)
+            raise ValidationError, (@validators[name][:message] || "configuration option for '#{name}' failed it's validation")
+          end
+        end
+      end
 
       def define_option_accessor name
         define_singleton_method(name) do |*args, &block|
@@ -48,6 +65,9 @@ module Statistrano
           end
         end
         define_singleton_method("#{name}=") { |arg| @options[name] = arg }
+      end
+
+      class ValidationError < StandardError
       end
 
   end